pyproject.toml
[project]
name = "bible-embeddings"
version = "0.1.0"
description = "Script to chunk Bible CSV by chapter and book, then upload embeddings to Neon database"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.2",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "marimo>=0.14.10",
    "plotly>=6.2.0",
    "networkx>=2.8.0",
    "neo4j>=5.19.0",
    "rapidfuzz>=3.9.4",
    "sentence-transformers>=2.7.0",
    "annoy>=1.17.3",
    "tqdm>=4.66.4",
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "pg8000>=1.31.2",
]

[project.optional-dependencies]
analysis = [
    "umap-learn>=0.5.0",     # Alternative to t-SNE for dimensionality reduction
    "hdbscan>=0.8.0",        # Advanced clustering algorithms
]
web = [
    "flask>=2.0.0",          # For serving interactive visualizations
    "jupyter>=1.0.0",        # For notebook-style analysis
]
dev = [
    "pytest>=7.0.0",         # Testing framework
    "black>=23.0.0",         # Code formatting
    "isort>=5.0.0",          # Import sorting
    "mypy>=1.0.0",           # Type checking
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "isort>=5.0.0",
    "mypy>=1.0.0",
] 
